requirements.txt
Flask==3.0.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
urllib3==2.1.0
python-dotenv==1.0.0

